classification/requirements.txt
numpy
Pillow
matplotlib
jpeg2dct
sklearn
PyTurboJPEG
tensorboardX
opencv-python
lmdb
pyarrow
scikit-image
